requirements.txt
einops==0.8.0
fairscale==0.4.13
fsspec==2024.3.1
ftfy==6.2.3
matplotlib==3.8.4
mmcv==2.2.0
mmengine==0.10.4
mmsegmentation==1.2.2
numpy==1.26.4
opencv-python==4.7.0
opencv_python_headless==4.8.0.76
openpyxl==3.1.5
regex==2024.9.11
safetensors==0.4.5
timm==1.0.9
torch==2.4.0
torchvision==0.19.0
tqdm==4.65.2
transformers==4.44.2